created local pre-commit hooks
- check docstrings for rapidfuzz funtions
- rendering readme
