+++ .pre-commit-config.yaml
@@ -46,3 +46,15 @@ repos:
     hooks:
       - id: flake8
         args: ["--max-line-length", "88"]
+  - repo: local
+    hooks:
+      - id: docstrings
+        name: checking docstrings
+        entry: poetry run python docs/docstrings.py
+        language: system
+        files: poetry.lock|pdfuzz.json
+      - id: readme
+        name: rendering readme
+        entry: poetry run python docs/readme.py
+        language: system
+        files: docs/README.md.jinja2
