+++ docs/__init__.py
@@ -1,0 +1,9 @@
+from .docstrings import main as check
+from .documentation import main as documentation
+from .readme import main as readme
+
+__all__ = [
+    "check",
+    "readme",
+    "documentation",
+]
